Restructure to deploy on dedicated server
vue-cli-service is no longer used as a server, but is still used to
"build" the app. A simple Express app has been implemented to serve
static resources from the build ('dist') directory, and is exported for
consumption by the root server app.

The MySQL server has not yet been installed on the server and so the
/data endpoint is unimplemented, resulting in an error when it is
requested (fetched) on page load.